Introduce persistent_id for persistent connections
When using `predis` extension, the underlying library will re-use
connections to same host/port in case of persistent connections. This
becomes an issue when the user's application uses the same host of Redis
but has different databases, meaning that this library will write to
whichever database is previously used by other parts of the code.

Using a persistent id forces the library to create a new connection
that will not conflict with any other connections being used by the
application.